requirements.txt
torch>=2.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
google-generativeai>=0.5.0
